RS_Checkpoint_2/auth-service/requirements.txt
fastapi
uvicorn
python-jose
python-dotenv
requests